Add files via upload
Project 2 uploaded here is a rough draft regarding my hypothesis on the AGTR1 gene in humans and mice. The file contains my scientific question, hypothesis, and explanation of the gene in both mice and humans. The scientific question asked in Project 2 is the angiotensin II receptor type 1 of a house mice more genetically expressed than a human, therefore explaining why there needs to be more blood regulation in mice due to their smaller bodies? The hypothesis that I concluded is if the angiotensin receptor II type 1 were to be more genetically expressed in mice that would then explain that mice need more blood regulation than humans due to their higher heart rate and respiratory rate because of their body size.
The mousetype1a and mousetype1b are fasta files containing the DNA sequence of the AGTR1 gene of a mouse. Meanwhile, humantype1 is a fasta file containing the DNA sequence of the AGTR1 gene of a human. All the fasta files are used in Project 2 to compare the AGTR1 gene in humans and mice so I can find the conclusion to the scientific question.